# apps/grc-platform/backend/openapi.yaml
openapi: 3.0.3
info:
  title: GRC Platform Backend API
  description: Public-facing API for the GRC Platform (Audit Hub and Risk Hub). Authenticates
    callers via a Bearer JWT issued by Asgardeo and validated against the issuer's JWKS
    (RS256, iss/aud checked). Most data access is proxied to the internal Compliance
    Entity service; file uploads are proxied through this backend to Azure Blob Storage
    rather than issuing direct SAS URLs to clients.
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Local development
...
security:
- bearerAuth: []
paths:
...
  /api/v1/audits/{id}/controls/{controlId}/evidence/submit:
    post:
      tags:
      - Audit Evidence
      summary: Submit an evidence round
      description: Records the accumulated uploaded files as a new evidence submission round and advances the control to EVIDENCE_INTERNAL_REVIEW.
      operationId: submitEvidence
      parameters:
      - $ref: '#/components/parameters/AuditId'
      - $ref: '#/components/parameters/ControlId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitEvidenceRequest'
      responses:
        '201':
          description: Evidence submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvidence'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/audits/{id}/controls/{controlId}/evidence/review:
    post:
      tags:
      - Audit Evidence
      summary: Internal-reviewer decision on a submitted evidence round
      description: Approve advances the control to EVIDENCE_UNDER_VALIDATION; reject sends it back to EVIDENCE_PENDING.
      operationId: reviewEvidence
      parameters:
      - $ref: '#/components/parameters/AuditId'
      - $ref: '#/components/parameters/ControlId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDecisionRequest'
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Asgardeo-issued JWT, validated by signature (RS256 via JWKS), issuer, and audience. Sent as `Authorization Bearer <token>`. Not required for GET /health.
    evidencePortalClientCredentials:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Client-credentials JWT for the Evidence Portal machine ingress
        (`/api/v1/evidence-portal/*`). Same Asgardeo org, issuer and JWKS as
        bearerAuth, but its `aud` is PORTAL_AUTH_AUDIENCE (the portal backend's
        consumer key) and its `sub` must be listed in PORTAL_CLIENTS. No grant
        load, no external-auditor guard — authorization is the token-derived
        team scope.
  parameters:
    AuditId:
      name: id
      in: path
      required: true
      description: Audit ID.
      schema:
        type: integer
    ControlId:
      name: controlId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller lacks the privilege required for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The resource is not in a state that allows this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: The request is well-formed but refers to something that could not be verified (e.g. an unrecognised employee email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Uploaded file exceeds the 25 MiB limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: Standard JSON error envelope returned on every non-2xx response.
      properties:
        message:
          type: string
    StatusResponse:
      type: object
      description: Generic status-change acknowledgement.
      properties:
        status:
          type: string
...
    AuditEvidenceFile:
      type: object
      properties:
        id:
          type: integer
        evidenceId:
          type: integer
        fileName:
          type: string
        filePath:
          type: string
        fileType:
          type: string
          nullable: true
        fileSize:
          type: integer
          format: int64
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        readUrl:
          type: string
          nullable: true
          description: Backend proxy download URL, computed at list time.
    AuditEvidence:
      type: object
      description: One evidence submission round for a control.
      properties:
        id:
          type: integer
        controlId:
          type: integer
        status:
          type: string
        folderPath:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvidenceFile'
        attestation:
          type: string
          description: Written justification for a round with no files. Empty for ordinary rounds.
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
...
    EvidenceFileRef:
      type: object
      description: Identifies one already-uploaded blob by its stored blob name and original file name.
      properties:
        blobName:
          type: string
        fileName:
          type: string
    SubmitEvidenceRequest:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceFileRef'
        attestation:
          type: string
          description: Fileless-completion justification. Only honored when files is empty AND the caller holds ManageControls.
        skipAiValidation:
          type: boolean
          description: Opts this one submission out of AI validation. Decided fresh per submission; never persisted.
...
    ReviewDecisionRequest:
      type: object
      required:
      - decision
      properties:
        decision:
          type: string
          enum:
          - APPROVE
          - REJECT
        comment:
          type: string
          nullable: true
...